Move to new template and clear content
Doing this for overall improvement in exposition. Staff was added as I
was learning it, and there was some duplication between the introduction
and the first chapter - better now to put everything back in write it
well once and for all.

Also a sick new style that took ages to cultivate.